refactor(web-kit): dedupe toErrorMessage into shared module [#1418]
3 SPA (admin-console / application-admin-console / participant-portal) の
`src/lib/error-message.ts` に **byte-identical** で重複していた `toErrorMessage`
(unknown throw 値 → 文字列正規化) を `@tenkacloud/web-kit` の単一 module に集約する。
usePolling (#1418) / buildLocaleUtility (#1626) と同じ DRY/SOLID 集約パターン。

- packages/web-kit: `error-message.ts` (canonical) + index export + 単体 test。
- 17 import site を `@tenkacloud/web-kit` import へ差し替え、 3 SPA の重複 source + test を削除。
- 共有 util は web-kit 側で 100% test。 各 SPA は import 先変更のみで挙動不変。

#1623 (同 #1431 epic) が commit し漏れた capacity-pressure runbook の生成 HTML も同梱。

## Regression analysis
- `toErrorMessage` の実装は 3 SPA で完全一致だったため、 集約で挙動は一切変わらない
  (`err instanceof Error ? err.message : String(err)`)。
- TenantDetail.test は旧 `../src/lib/error-message` を `String(e)` で mock し "Error: list boom"
  を assert していた。 削除した module の mock は無効なので mock を外し、 実物 (= raw message
  "list boom") を assert するよう更新。 他の test は error-message を mock していない。
- coverage gate: 全 10 gated workspace 100% (= web-kit に新規 test、 3 SPA は import 先変更のみ)。
- 全 workspace test green (web-kit 41 / admin-console 265 / app-admin 902 / portal 592)、
  tsc / biome / lint / harness (error 0) / cdk synth / 全 gate pass。

## Physical impact
- NO-OP on CFn / deployed artifacts。 frontend source-only の DRY 集約。 各 SPA dist は
  toErrorMessage が web-kit 由来になるだけで bundle 挙動は同一 → UPDATE (実質 NO-OP)。
- docs: capacity-pressure*.html 生成物 → CREATE (docs のみ)。

Relates #1418
