+++ apps/application-admin-console/src/hooks/useEventDetail.ts
@@ -1,7 +1,7 @@
+import { toErrorMessage } from "@tenkacloud/web-kit";
 import { useCallback, useEffect, useState } from "react";
 import type { ApiClient } from "../api/client";
 import { type EventDetail, getEvent } from "../api/events-client";
-import { toErrorMessage } from "../lib/error-message";
 
 export function useEventDetail(args: {
   readonly apiClient: ApiClient | null;
