+++ apps/application-admin-console/src/pages/EventCreate.tsx
@@ -5,6 +5,7 @@ import Button from "@cloudscape-design/components/button";
 import Form from "@cloudscape-design/components/form";
 import Header from "@cloudscape-design/components/header";
 import SpaceBetween from "@cloudscape-design/components/space-between";
+import { toErrorMessage } from "@tenkacloud/web-kit";
 import { useCallback, useMemo, useState } from "react";
 import { useNavigate } from "react-router";
 import { useApiClient } from "../api/client";
@@ -14,7 +15,6 @@ import { DEFAULT_AWS_REGION } from "../data/aws-regions";
 import { listProblemSummaries } from "../data/problems";
 import { useT } from "../i18n";
 import { filterVerifiedAccounts } from "../lib/competitor-accounts-filter";
-import { toErrorMessage } from "../lib/error-message";
 import { EventCreateAccountsAlerts } from "./event-create/EventCreateAccountsAlerts";
 import { EventCreateBasicInfoSection } from "./event-create/EventCreateBasicInfoSection";
 import { EventCreateDeployPromptModal } from "./event-create/EventCreateDeployPromptModal";
